Шифрование пароля - всесто создания объекта PasswordEncoder завел бин - убрал имплементацию UserDetailService у класа UserServiceImpl - создал класс UserDetailServiceImpl, имплементирующий UserDetailService - поправил SecurityConfig, теперь в бин userDetailService инжектится корректный класс
