+++ src/main/java/ru/javamentor/config/SecurityConfig.java
@@ -24,13 +24,16 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
     private UserDetailsService userDetailsService;
 
     @Autowired
-    public SecurityConfig(@Qualifier("userServiceImpl") UserDetailsService userDetailsService) {
+    private PasswordEncoder passwordEncoder;
+
+    @Autowired
+    public SecurityConfig(@Qualifier("userDetailServiceImpl") UserDetailsService userDetailsService) {
         this.userDetailsService = userDetailsService;
     }
 
     @Override
     protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.userDetailsService(userDetailsService).passwordEncoder(new BCryptPasswordEncoder());
+        auth.userDetailsService(userDetailsService).passwordEncoder(passwordEncoder);
 /*        auth.inMemoryAuthentication()
                 .withUser("user").password("user").roles("USER").and()
                 .withUser("admin").password("admin").roles("ADMIN");*/
@@ -59,5 +62,10 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .antMatchers("/api/admin/**").hasAuthority("ADMIN")
                 .antMatchers("/api/user/*").hasAnyAuthority("USER", "ADMIN")
                 .anyRequest().authenticated();
+    }
+
+    @Bean
+    public PasswordEncoder getBCryptPasswordEncoder() {
+        return new BCryptPasswordEncoder();
     }
 }
