+++ src/main/java/ru/javamentor/controller/PageController.java
@@ -3,6 +3,7 @@ package ru.javamentor.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -22,7 +23,7 @@ public class PageController {
     public final TopicService topicService;
 
     @Autowired
-    private UserServiceImpl userService;
+    private UserDetailsService userService;
 
     @Autowired
     public PageController(TopicService topicService) {
